# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/schedule_organizations:
    get:
      tags:
        - ScheduleOrganizations
      summary: List ScheduleOrganizations
      description: Get a paginated list of ScheduleOrganizations with optional filtering and sorting
      operationId: listScheduleOrganizations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleOrganizationList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ScheduleOrganizations
      summary: Create ScheduleOrganization
      description: Create a new ScheduleOrganization
      operationId: createScheduleOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleOrganizationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleOrganization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/schedule_weekdays:
    get:
      tags:
        - ScheduleWeekdays
      summary: List ScheduleWeekdays
      description: Get a paginated list of ScheduleWeekdays with optional filtering and sorting
      operationId: listScheduleWeekdays
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleWeekdayList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - ScheduleWeekdays
      summary: Create ScheduleWeekday
      description: Create a new ScheduleWeekday
      operationId: createScheduleWeekday
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleWeekdayRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleWeekday'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    ScheduleOrganization:
      type: object
      required:
        - id
        - order_number
        - time_in
        - time_out
        - metadata
      properties:
        id:
          type: string
          format: uuid
        structure_id:
          nullable: true
          type: string
          format: uuid
        name:
          nullable: true
          type: string
        order_number:
          type: integer
          format: int32
        start_date:
          nullable: true
          type: string
          format: date
        end_date:
          nullable: true
          type: string
          format: date
        time_in:
          type: string
          format: time
        time_out:
          type: string
          format: time
        metadata:
          type: object
...
    CreateScheduleOrganizationRequest:
      type: object
      required:
        - order_number
        - time_in
        - time_out
        - metadata
      properties:
        structure_id:
          nullable: true
          type: string
          format: uuid
        name:
          nullable: true
          type: string
        order_number:
          type: integer
          format: int32
        start_date:
          nullable: true
          type: string
          format: date
        end_date:
          nullable: true
          type: string
          format: date
        time_in:
          type: string
          format: time
        time_out:
          type: string
          format: time
        metadata:
          type: object
...
    ScheduleOrganizationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleOrganization'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ScheduleWeekday:
      type: object
      required:
        - id
        - schedule_id
        - sun
        - mon
        - tue
        - wed
        - thu
        - fri
        - sat
        - metadata
      properties:
        id:
          type: string
          format: uuid
        schedule_id:
          type: string
          format: uuid
        sun:
          type: string
        mon:
          type: string
        tue:
          type: string
        wed:
          type: string
        thu:
          type: string
        fri:
          type: string
        sat:
          type: string
        metadata:
          type: object
...
    CreateScheduleWeekdayRequest:
      type: object
      required:
        - schedule_id
        - sun
        - mon
        - tue
        - wed
        - thu
        - fri
        - sat
        - metadata
      properties:
        schedule_id:
          type: string
          format: uuid
        sun:
          type: string
        mon:
          type: string
        tue:
          type: string
        wed:
          type: string
        thu:
          type: string
        fri:
          type: string
        sat:
          type: string
        metadata:
          type: object
...
    ScheduleWeekdayList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleWeekday'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []